new `wikidata.sh` script
As a step towards doing something about #3, ie, feeding the e-mail
addresses info to wikidata, where it would be most useful, a new
`wikidata.sh` script was created. It is meant to be ran after the
`get-emails.sh` script (ie, when an `emails` file exists), and it will
generate a CSV file (`wd.csv`) with the wikidata items for each MEP
for which we know an e-mail address, and the e-mail address (if any)
that wikidata knows about them. With both `emails` and `wd.csv`
generated, it would be easy to generate a quickstatements input file
to feed all the missing email addresses into wikidata.